ingest/requirements.txt
procyclingstats==0.2.3
supabase==2.10.0
python-dotenv==1.0.1
urllib3<2
httpx==0.27.2


